Move "game" classes to base.py
